fix(avatar): demo规范与调整
