+++ src/packages/avatar/demos/taro/demo1.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import { Avatar } from '@nutui/nutui-react-taro'
+
+const Demo1 = () => {
+  return (
+    <>
+      <Avatar
+        size="large"
+        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
+      />
+      <Avatar
+        size="normal"
+        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
+      />
+      <Avatar
+        size="small"
+        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
+      />
+    </>
+  )
+}
+export default Demo1
