+++ src/packages/avatar/doc.en-US.md
@@ -18,30 +18,7 @@ Support three sizes：small、normal、large
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar, Cell } from '@nutui/nutui-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar
-        size="large"
-        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
-      />
-      <Avatar
-        size="normal"
-        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
-      />
-      <Avatar
-        size="small"
-        src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
-      />
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -51,21 +28,7 @@ Support two shapes：square、round
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar icon={<User />} shape="square" />
-      <Avatar icon={<User />} shape="round" />
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo2.tsx'></CodeBlock>
 
 :::
 
@@ -75,22 +38,7 @@ Support three types：picture、icon、letter
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png" />
-      <Avatar icon={<User />} />
-      <Avatar>N</Avatar>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo3.tsx'></CodeBlock>
 
 :::
 
@@ -100,26 +48,7 @@ Icon and letter types can have custom colors and background colors
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar
-        className="demo-avatar"
-        color="#fff"
-        background="#FA2C19"
-        icon={<User />}
-      />
-      <Avatar color="rgb(245, 106, 0)" background="rgb(253, 227, 207)">U</Avatar>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo4.tsx'></CodeBlock>
 
 :::
 
@@ -127,25 +56,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar, Badge } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Badge value="8">
-        <Avatar icon={<User />} shape="square" />
-      </Badge>
-      <Badge dot>
-        <Avatar icon={<User />} shape="square" />
-      </Badge>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo5.tsx'></CodeBlock>
 
 :::
 
@@ -153,35 +64,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar.Group gap="-4">
-        <Avatar src="https://img12.360buyimg.com/imagetools/jfs/t1/196430/38/8105/14329/60c806a4Ed506298a/e6de9fb7b8490f38.png" />
-        <Avatar icon={<User />} />
-        <Avatar color="rgb(245, 106, 0)" background="rgb(253, 227, 207)">
-          U
-        </Avatar>
-      </Avatar.Group>
-
-      <Avatar.Group max="3" maxColor="#fff" maxBackground="#498ff2">
-        <Avatar src="https://img12.360buyimg.com/imagetools/jfs/t1/196430/38/8105/14329/60c806a4Ed506298a/e6de9fb7b8490f38.png" />
-        <Avatar icon={<User />} />
-        <Avatar color="rgb(245, 106, 0)" background="rgb(253, 227, 207)">
-          U
-        </Avatar>
-        <Avatar icon={<User />} />
-      </Avatar.Group>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo6.tsx'></CodeBlock>
 
 :::
 
@@ -189,27 +72,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Avatar.Group max="3" level="right" maxContent="...">
-        <Avatar src="https://img12.360buyimg.com/imagetools/jfs/t1/196430/38/8105/14329/60c806a4Ed506298a/e6de9fb7b8490f38.png" />
-        <Avatar icon={<User />} />
-        <Avatar color="rgb(245, 106, 0)" background="rgb(253, 227, 207)">
-          U
-        </Avatar>
-        <Avatar icon={<User />} />
-      </Avatar.Group>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo7.tsx'></CodeBlock>
 
 :::
 
@@ -217,23 +80,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar, Toast } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  const activeAvatar = () => {
-    Toast.show('触发点击头像')
-  }
-  return (
-    <>
-      <Avatar icon={<User />} onClick={activeAvatar} />
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo8.tsx'></CodeBlock>
 
 :::
 
@@ -241,30 +88,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Avatar, Toast, Cell } from '@nutui/nutui-react';
-import { User } from '@nutui/icons-react';
-
-const App = () => {
-  return (
-    <>
-      <Cell>
-        <Avatar
-          size="normal"
-          style={{ marginRight: 0 }}
-          src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
-        />
-        <div style={{ marginLeft: '10px' }}>
-          <div style={{ fontSize: '16px' }}>Title</div>
-          <div style={{ fontSize: '12px' }}>Description</div>
-        </div>
-      </Cell>
-    </>
-  )
-}
-export default App;
-```
+<CodeBlock src='h5/demo9.tsx'></CodeBlock>
 
 :::
